+++ src/core/session/att/att.cpp
@@ -41,11 +41,9 @@ void AttTracer::AddPendingSignals(
   sessions_pending_signals_.at(writer_id).emplace_back(att_pending_signal_t{
       kernel_object, original_completion_signal, new_completion_signal, session_id_, buffer_id,
       profile, kernel_properties, thread_id, queue_index});
-  std::atomic_thread_fence(std::memory_order_release);
 }
 
 const std::vector<att_pending_signal_t>& AttTracer::GetPendingSignals(uint32_t writer_id) {
-  std::atomic_thread_fence(std::memory_order_acquire);
   std::lock_guard<std::mutex> lock(sessions_pending_signals_lock_);
   assert(sessions_pending_signals_.find(writer_id) != sessions_pending_signals_.end() &&
          "writer_id is not found in the pending_signals");